+++ creme/feature_extraction/__init__.py
@@ -2,6 +2,7 @@
 A module for extracting features from streaming data.
 """
 from .gb import GroupBy
+from .gb import TargetGroupBy
 from .target_encoding import TargetEncoder
 from .vectorize import CountVectorizer
 from .vectorize import TFIDFVectorizer
@@ -11,5 +12,6 @@ __all__ = [
     'CountVectorizer',
     'GroupBy',
     'TargetEncoder',
+    'TargetGroupBy',
     'TFIDFVectorizer'
 ]