+++ creme/stats/rolling_mean.py
@@ -36,7 +36,7 @@ class RollingMean(rolling_sum.RollingSum):
 
     @property
     def name(self):
-        return 'rolling_{self.window_size}_mean'
+        return f'rolling_{self.window_size}_mean'
 
     def get(self):
         return super().get() / len(self) if len(self) > 0 else 0